backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
dateparser
pydantic
chromadb
langchain
langchain-google-genai
langchain-huggingface
langchain-chroma
sentence-transformers
pypdf
python-dotenv
httpx
langchain-groq
langchain-openai
openai
langgraph
